Reorganize Loadable into internal module
